refactor(whatsrust): extract contacts and communities queries
